Project edit pre-work setup
