+++ api/src/app.ts
@@ -76,6 +76,8 @@ initialize({
       defaultLog.error({ label: 'errorMiddleware', message: 'unexpected error', error });
     }
 
+    console.log(error);
+    console.log(Object.keys(error));
     res.status(error.status || 500).json(error);
   }
 });
