+++ api/src/models/project-create.test.ts
@@ -1,17 +1,13 @@
 import { expect } from 'chai';
 import { describe } from 'mocha';
 import {
-  PostProjectData,
+  PostIUCNData,
   PostLocationData,
   PostObjectivesData,
-  PostIUCNData,
-  GetIUCNClassificationData,
-  GetFundingData,
-  PostSpeciesData,
-  GetSpeciesData,
   PostPartnershipsData,
-  GetPartnershipsData
-} from './project';
+  PostProjectData,
+  PostSpeciesData
+} from './project-create';
 
 describe('PostProjectData', () => {
   describe('No values provided', () => {
@@ -214,156 +210,6 @@ describe('PostSpeciesData', () => {
   });
 });
 
-describe('GetSpeciesData', () => {
-  describe('No values provided', () => {
-    let data: GetSpeciesData;
-
-    before(() => {
-      data = new GetSpeciesData((null as unknown) as any[], []);
-    });
-
-    it('sets focal_species', function () {
-      expect(data.focal_species).to.eql([]);
-    });
-
-    it('sets ancillary_species', function () {
-      expect(data.ancillary_species).to.eql([]);
-    });
-  });
-
-  describe('focal species values provided', () => {
-    let data: GetSpeciesData;
-
-    const focal_species = [{ name: 'species 1' }, { name: 'species 2' }];
-    const ancillary_species: any[] = [];
-
-    before(() => {
-      data = new GetSpeciesData(focal_species, ancillary_species);
-    });
-
-    it('sets focal_species', function () {
-      expect(data.focal_species).to.eql(['species 1', 'species 2']);
-    });
-
-    it('sets ancillary_species', function () {
-      expect(data.ancillary_species).to.eql([]);
-    });
-  });
-
-  describe('ancillary species values provided', () => {
-    let data: GetSpeciesData;
-
-    const focal_species = (null as unknown) as any[];
-    const ancillary_species = [{ name: 'species 3' }, { name: 'species 4' }];
-
-    before(() => {
-      data = new GetSpeciesData(focal_species, ancillary_species);
-    });
-
-    it('sets focal_species', function () {
-      expect(data.focal_species).to.eql([]);
-    });
-
-    it('sets ancillary_species', function () {
-      expect(data.ancillary_species).to.eql(['species 3', 'species 4']);
-    });
-  });
-
-  describe('All values provided', () => {
-    let data: GetSpeciesData;
-
-    const focal_species = [{ name: 'species 1' }, { name: 'species 2' }];
-    const ancillary_species = [{ name: 'species 3' }];
-
-    before(() => {
-      data = new GetSpeciesData(focal_species, ancillary_species);
-    });
-
-    it('sets focal_species', function () {
-      expect(data.focal_species).to.eql(['species 1', 'species 2']);
-    });
-
-    it('sets ancillary_species', function () {
-      expect(data.ancillary_species).to.eql(['species 3']);
-    });
-  });
-});
-
-describe('GetPartnershipsData', () => {
-  describe('No values provided', () => {
-    let data: GetPartnershipsData;
-
-    before(() => {
-      data = new GetPartnershipsData([], []);
-    });
-
-    it('sets indigenous_partnerships', function () {
-      expect(data.indigenous_partnerships).to.eql([]);
-    });
-
-    it('sets stakeholder_partnerships', function () {
-      expect(data.stakeholder_partnerships).to.eql([]);
-    });
-  });
-
-  describe('indigenous_partnerships values provided', () => {
-    let data: GetPartnershipsData;
-
-    const indigenous_partnerships = [{ name: 'partner 1' }, { name: 'partner 2' }];
-    const stakeholder_partnerships: string[] = [];
-
-    before(() => {
-      data = new GetPartnershipsData(indigenous_partnerships, stakeholder_partnerships);
-    });
-
-    it('sets indigenous_partnerships', function () {
-      expect(data.indigenous_partnerships).to.eql(['partner 1', 'partner 2']);
-    });
-
-    it('sets stakeholder_partnerships', function () {
-      expect(data.stakeholder_partnerships).to.eql([]);
-    });
-  });
-
-  describe('stakeholder_partnerships values provided', () => {
-    let data: GetPartnershipsData;
-
-    const indigenous_partnerships: string[] = [];
-    const stakeholder_partnerships = [{ name: 'partner 1' }, { name: 'partner 2' }];
-
-    before(() => {
-      data = new GetPartnershipsData(indigenous_partnerships, stakeholder_partnerships);
-    });
-
-    it('sets indigenous_partnerships', function () {
-      expect(data.indigenous_partnerships).to.eql([]);
-    });
-
-    it('sets stakeholder_partnerships', function () {
-      expect(data.stakeholder_partnerships).to.eql(['partner 1', 'partner 2']);
-    });
-  });
-
-  describe('All values provided', () => {
-    let data: GetPartnershipsData;
-
-    const indigenous_partnerships = [{ name: 'partner 1' }, { name: 'partner 2' }];
-    const stakeholder_partnerships = [{ name: 'partner 3' }, { name: 'partner 4' }];
-
-    before(() => {
-      data = new GetPartnershipsData(indigenous_partnerships, stakeholder_partnerships);
-    });
-
-    it('sets indigenous_partnerships', function () {
-      expect(data.indigenous_partnerships).to.eql(['partner 1', 'partner 2']);
-    });
-
-    it('sets stakeholder_partnerships', function () {
-      expect(data.stakeholder_partnerships).to.eql(['partner 3', 'partner 4']);
-    });
-  });
-});
-
 describe('PostIUCNData', () => {
   describe('No values provided', () => {
     let projectIUCNData: PostIUCNData;
@@ -396,83 +242,6 @@ describe('PostIUCNData', () => {
 
     it('sets classification details', function () {
       expect(projectIUCNData.classificationDetails).to.eql(obj.classificationDetails);
-    });
-  });
-});
-
-describe('GetIUCNClassificationData', () => {
-  describe('No values provided', () => {
-    let iucnClassificationData: GetIUCNClassificationData;
-
-    before(() => {
-      iucnClassificationData = new GetIUCNClassificationData([]);
-    });
-
-    it('sets classification details', function () {
-      expect(iucnClassificationData.classificationDetails).to.eql([]);
-    });
-  });
-
-  describe('All values provided', () => {
-    let iucnClassificationData: GetIUCNClassificationData;
-
-    const iucnClassificationDataObj = [
-      {
-        classification: 'class',
-        subclassification1: 'subclass1',
-        subclassification2: 'subclass2'
-      }
-    ];
-
-    before(() => {
-      iucnClassificationData = new GetIUCNClassificationData(iucnClassificationDataObj);
-    });
-
-    it('sets classification details', function () {
-      expect(iucnClassificationData.classificationDetails).to.eql([
-        {
-          classification: 'class',
-          subClassification1: 'subclass1',
-          subClassification2: 'subclass2'
-        }
-      ]);
-    });
-  });
-});
-
-describe('GetFundingData', () => {
-  describe('No values provided', () => {
-    let fundingData: GetFundingData;
-
-    before(() => {
-      fundingData = new GetFundingData([]);
-    });
-
-    it('sets funding agencies', function () {
-      expect(fundingData.fundingAgencies).to.eql([]);
-    });
-  });
-
-  describe('All values provided', () => {
-    let fundingData: GetFundingData;
-
-    const fundingDataObj = [
-      {
-        agency_id: '123',
-        agency_name: 'Agency name',
-        investment_action_category: 'investment',
-        start_date: '01/01/2020',
-        end_date: '01/01/2021',
-        funding_amount: 123
-      }
-    ];
-
-    before(() => {
-      fundingData = new GetFundingData(fundingDataObj);
-    });
-
-    it('sets funding agencies', function () {
-      expect(fundingData.fundingAgencies).to.eql(fundingDataObj);
     });
   });
 });
