+++ api/src/models/project-update.ts
@@ -1,0 +1,70 @@
+import { Feature } from 'geojson';
+import { getLogger } from '../utils/logger';
+
+const defaultLog = getLogger('models/project-update');
+
+export class PutProjectData {
+  name: string;
+  type: number;
+  start_date: string;
+  end_date: string;
+  revision_count: number;
+
+  constructor(obj?: any) {
+    defaultLog.debug({ label: 'PutProjectData', message: 'params', obj });
+
+    this.name = obj?.project_name || null;
+    this.type = obj?.project_type || null;
+    this.start_date = obj?.start_date || null;
+    this.end_date = obj?.end_date || null;
+    this.revision_count = obj?.revision_count || null;
+  }
+}
+
+export class PutLocationData {
+  location_description: string;
+  geometry: Feature[];
+  revision_count: number;
+
+  constructor(obj?: any) {
+    defaultLog.debug({ label: 'PutLocationData', message: 'params', obj });
+
+    this.location_description = (obj && obj.location_description) || null;
+    this.geometry = (obj?.geometry?.length && obj.geometry) || [];
+    this.revision_count = obj?.revision_count || null;
+  }
+}
+
+export class PutObjectivesData {
+  objectives: string;
+  caveats: string;
+  revision_count: number;
+
+  constructor(obj?: any) {
+    defaultLog.debug({ label: 'PutObjectivesData', message: 'params', obj });
+
+    this.objectives = obj?.objectives || '';
+    this.caveats = obj?.caveats || null;
+    this.revision_count = obj?.revision_count || null;
+  }
+}
+
+export class PutCoordinatorData {
+  first_name: string;
+  last_name: string;
+  email_address: string;
+  coordinator_agency: string;
+  share_contact_details: boolean;
+  revision_count: number;
+
+  constructor(obj?: any) {
+    defaultLog.debug({ label: 'PutCoordinatorData', message: 'params', obj });
+
+    this.first_name = obj?.first_name || null;
+    this.last_name = obj?.last_name || null;
+    this.email_address = obj?.email_address || null;
+    this.coordinator_agency = obj?.coordinator_agency || null;
+    this.share_contact_details = (obj?.share_contact_details === 'true' && true) || false;
+    this.revision_count = obj?.revision_count || null;
+  }
+}
