+++ api/src/paths/permit-no-sampling.ts
@@ -3,7 +3,7 @@ import { Operation } from 'express-openapi';
 import { WRITE_ROLES } from '../constants/roles';
 import { getDBConnection, IDBConnection } from '../database/db';
 import { IPostPermitNoSampling, PostPermitNoSamplingObject } from '../models/permit-no-sampling';
-import { PostCoordinatorData } from '../models/project';
+import { PostCoordinatorData } from '../models/project-create';
 import { permitNoSamplingPostBody, permitNoSamplingResponseBody } from '../openapi/schemas/permit-no-sampling';
 import { postPermitNoSamplingSQL } from '../queries/permit-no-sampling/permit-no-sampling-queries';
 import { getLogger } from '../utils/logger';
