+++ api/src/queries/permit-no-sampling/permit-no-sampling-queries.ts
@@ -1,6 +1,6 @@
 import { SQL, SQLStatement } from 'sql-template-strings';
 import { IPostPermitNoSampling } from '../../models/permit-no-sampling';
-import { PostCoordinatorData } from '../../models/project';
+import { PostCoordinatorData } from '../../models/project-create';
 import { getLogger } from '../../utils/logger';
 
 const defaultLog = getLogger('queries/permit-no-sampling/permit-no-sampling-queries');
