+++ api/src/queries/project/project-create-queries.test.ts
@@ -6,7 +6,7 @@ import {
   PostLocationData,
   PostObjectivesData,
   PostProjectData
-} from '../../models/project';
+} from '../../models/project-create';
 import {
   postAncillarySpeciesSQL,
   postFocalSpeciesSQL,
